app/requirement.txt
aiofiles==0.7.0
aiosqlite==0.17.0
asgiref==3.3.4
attrs==19.3.0
Automat==0.8.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.27
bcrypt==3.2.0
bidict==0.21.2
blinker==1.4
cachetools==4.2.2
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
cloud-init==20.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
crcmod==1.7
cryptography==2.8
cycler==0.10.0
databases==0.4.3
dbus-python==1.2.16
decorator==4.4.2
Deprecated==1.2.12
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
email-validator==1.1.3
entrypoints==0.3
fail2ban==0.11.1
fastapi==0.65.2
fastapi-socketio==0.0.6
fastapi-users==6.1.0
gevent==21.1.2
google-api-core==1.30.0
google-auth==1.31.0
google-cloud==0.34.0
google-cloud-vision==2.3.2
googleapis-common-protos==1.53.0
greenlet==1.1.0
grpcio==1.38.0
gunicorn==20.1.0
h11==0.12.0
h2==4.0.0
hpack==4.0.0
httpie==2.4.0
httplib2==0.14.0
httptools==0.2.0
Hypercorn==0.11.2
hyperframe==6.0.1
hyperlink==19.0.0
idna==2.8
imageio==2.9.0
importlib-metadata==1.5.0
imutils==0.5.4
incremental==16.10.1
isodate==0.6.0
Jinja2==2.10.1
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
makefun==1.11.3
MarkupSafe==1.1.0
matplotlib==3.4.2
more-itertools==4.2.0
msrest==0.6.21
netifaces==0.11.0
networkx==2.5.1
numpy==1.20.3
oauthlib==3.1.0
opencv-python==4.5.2.54
packaging==20.9
passlib==1.7.4
pexpect==4.6.0
Pillow==8.2.0
priority==1.3.0
proto-plus==1.18.1
protobuf==3.17.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
PyGObject==3.36.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyrsistent==0.15.5
pyserial==3.4
PySocks==1.7.1
pytesseract==0.3.7
python-apt==2.0.0+ubuntu0.20.4.3
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
python-engineio==4.2.0
python-multipart==0.0.5
python-socketio==5.3.0
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.7.2
scikit-image==0.18.1
scipy==1.7.0
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.0
SQLAlchemy==1.3.24
ssh-import-id==5.10
starlette==0.14.2
systemd-python==234
tesseract==0.1.3
tifffile==2021.6.14
toml==0.10.2
Twisted==18.9.0
typing-extensions==3.10.0.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.5
uvicorn==0.14.0
uvloop==0.15.2
wadllib==1.3.3
websocket==0.2.1
websocket-client==1.1.0
websocket-server==0.4
websockets==9.1
wincertstore==0.2
wrapt==1.12.1
wsproto==0.15.0
zipp==1.0.0
zope.event==4.5.0
zope.interface==4.7.1
